requirements.txt
anthropic>=0.39.0
requests>=2.31.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
sentence-transformers>=2.2.0
torch>=2.0.0

# Testing (
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality )
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0